fix(Burnsville): Founsd new API that's faster and returns json - Updated Burnville to use new endpoint
The endpoint does not include Stick and Puck events. This seems to be because it's an API maintained by the city and the Ice Center maintains its own calendar. The problem is that the Ice Arena calendar is so slow to load.

That said, I do know that the Ice Arena matches up stick and puck events with public skating events except for Sundays(no events on Saturdays).

So we'll just use the new API and the above logic to generate Stick and Puck events.